+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/model/BukkitChatFormat.kt
@@ -7,9 +7,9 @@ import dev.slne.surf.chat.api.type.ChatMessageType
 import dev.slne.surf.chat.bukkit.extension.LuckPermsExtension
 import dev.slne.surf.chat.bukkit.plugin
 import dev.slne.surf.chat.bukkit.util.components
-import dev.slne.surf.chat.bukkit.util.pluginConfig
-import dev.slne.surf.chat.bukkit.util.serverPlayers
-import dev.slne.surf.chat.bukkit.util.toPlainText
+import dev.slne.surf.chat.bukkit.util.utils.pluginConfig
+import dev.slne.surf.chat.bukkit.util.utils.serverPlayers
+import dev.slne.surf.chat.bukkit.util.utils.toPlainText
 import dev.slne.surf.chat.core.service.databaseService
 import dev.slne.surf.surfapi.core.api.messages.Colors
 import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
