+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/util/utils/chatchannel-util.kt
@@ -1,4 +1,4 @@
-package dev.slne.surf.chat.bukkit.util
+package dev.slne.surf.chat.bukkit.util.utils
 
 import dev.slne.surf.chat.api.model.ChannelModel
 import dev.slne.surf.chat.core.service.channelService
